+++ src/AdminSettings.API/Data/AdminSettingsDbContext.cs
@@ -6,9 +6,8 @@ namespace AdminSettings.Data;
 public class AdminSettingsDbContext : DbContext
 {
     public AdminSettingsDbContext(DbContextOptions<AdminSettingsDbContext> options)
-        : base(options)
-    {
-    }
+        : base(options) { }
+
     public DbSet<AuditLog> AuditLogs { get; set; }
-    
+    public DbSet<User> Users { get; set; }
 }
