+++ src/AdminSettings.API/Services/UserService.cs
@@ -1,4 +1,4 @@
-﻿
+﻿using System.Text.Json;
 using AdminSettings.Persistence.Entities;
 
 namespace AdminSettings.Services;
@@ -6,17 +6,19 @@ namespace AdminSettings.Services;
 public class UserService
 {
     private readonly HttpClient _httpClient;
-    
+
     public UserService(HttpClient httpClient)
     {
         _httpClient = httpClient;
     }
-    
-    public async Task<User> GetUserById(int id)
+
+    public async Task<User?> GetUserById(string userId)
     {
-        var response = await _httpClient.GetAsync($"api/user/{id}");
-        response.EnsureSuccessStatusCode();
-        var user = await response.Content.ReadFromJsonAsync<User>();
-        return user!;
+        var response = await _httpClient.GetAsync($"api/user/{userId}");
+        if (!response.IsSuccessStatusCode)
+            return null;
+
+        var json = await response.Content.ReadAsStringAsync();
+        return JsonSerializer.Deserialize<User>(json, new JsonSerializerOptions { PropertyNameCaseInsensitive = true });
     }
 }
