+++ src/AdminSettings.API/appsettings.json
@@ -1,7 +1,7 @@
 {
-    "ConnectionStrings": {
-  "DefaultConnection": "Server=;Database=;;MultipleActiveResultSets=true"
-},
+  "ConnectionStrings": {
+    "DefaultConnection": "Server=localhost;Database=db;User=root;Password=root"
+  },
   "Logging": {
     "LogLevel": {
       "Default": "Information",
